requirements.txt
flask==3.0.2
Pillow==10.2.0
qrcode==7.4.2
gunicorn==21.2.0 